fixed bugs
fixed bugs

+++ Tv-Show-Movie-Server/src/main/java/com/portfolio/TvMoviePicker/dao/JdbcMovieDao.java
@@ -163,7 +163,7 @@ public class JdbcMovieDao implements MovieDao {
         movie.setStreamingService(rs.getString("movie_streaming_service"));
         movie.setGenre1(rs.getString("movie_genre1"));
         movie.setGenre2(rs.getString("movie_genre2"));
-        movie.setImdbRating(rs.getInt("movie_IMBD_rating"));
+        movie.setImdbRating(rs.getBigDecimal("movie_IMBD_rating"));
         movie.setRtRating(rs.getInt("movie_RT_rating"));
         movie.setDescription(rs.getString("movie_description"));
         movie.setKeyword1(rs.getString("movie_keyword1"));
