+++ Tv-Show-Movie-Server/src/main/java/com/portfolio/TvMoviePicker/dao/JdbcTvShowDao.java
@@ -161,7 +161,7 @@ public class JdbcTvShowDao implements TvShowDao{
         tvShow.setStreamingService(rs.getString("tv_show_streaming_service"));
         tvShow.setGenre1(rs.getString("tv_show_genre1"));
         tvShow.setGenre2(rs.getString("tv_show_genre2"));
-        tvShow.setImdbRating(rs.getInt("tv_show_IMBD_rating"));
+        tvShow.setImdbRating(rs.getBigDecimal("tv_show_IMBD_rating"));
         tvShow.setRtRating(rs.getInt("tv_show_RT_rating"));
         tvShow.setDescription(rs.getString("tv_show_description"));
         tvShow.setKeyword1(rs.getString("tv_show_keyword1"));
